+++ packages/typespec-rust/src/codegen/cargotoml.ts
@@ -22,7 +22,8 @@ export function emitCargoToml(crate: rust.Crate): string {
     content += '\n[dependencies]\n';
     for (const dependency of crate.dependencies) {
       // dependency versions are managed by the workspace's Cargo.toml file
-      content += `${dependency.name} = { workspace = true }\n`;
+      const features = dependency.features.length > 0 ? `, features = [${dependency.features.sort().map(f => `"${f}"`).join(', ')}]` : '';
+      content += `${dependency.name} = { workspace = true${features} }\n`;
     }
   }
   return content;