+++ packages/typespec-rust/src/codemodel/crate.ts
@@ -44,6 +44,9 @@ export type ServiceType = 'azure-arm' | 'data-plane';
 export interface CrateDependency {
   /** the name of the Crate */
   name: string;
+
+  /** the features to enable for the Crate */
+  features: Array<string>;
 }
 
 /** Module is an entry in a mod.rs file */
@@ -76,6 +79,8 @@ export class Crate implements Crate {
   addDependency(dependency: CrateDependency): void {
     for (const dep of this.dependencies) {
       if (dep.name === dependency.name) {
+        // merge in any features
+        dep.features = dep.features.concat(dependency.features.filter(item => !dep.features.includes(item)));
         return;
       }
     }
@@ -115,8 +120,9 @@ export class Crate implements Crate {
 }
 
 export class CrateDependency implements CrateDependency {
-  constructor(name: string) {
+  constructor(name: string, features = new Array<string>()) {
     this.name = name;
+    this.features = features;
   }
 }